update: refactor helper methods from authOptions to web3 service
